+++ tests/RoomTest.php
@@ -6,7 +6,6 @@ use PHPUnit\Framework\TestCase;
 use PfinalClub\AsyncioGamekit\Room;
 use PfinalClub\AsyncioGamekit\Player;
 use PfinalClub\AsyncioGamekit\Tests\Helpers\TestRoom;
-use Generator;
 
 class RoomTest extends TestCase
 {